+++ app/js/preloads/actions.js
@@ -148,6 +148,18 @@ const writeInputToFile = filename => {
 
 const print = () => document.execCommand("print")
 
+const installFirefoxExtension = () => {
+    const link = [...document.querySelectorAll("a")].find(
+        a => a.href?.endsWith(".xpi"))?.href
+    const extension = link?.replace(/.*\/(\d{7,10}).*/g, "$1")
+    if (link && extension) {
+        ipcRenderer.send("install-extension", link, extension, "xpi")
+    } else {
+        ipcRenderer.sendToHost("notify",
+            "No extensions found on this page", "warn")
+    }
+}
+
 const functions = {
     blur,
     increasePageNumber,
@@ -168,7 +180,8 @@ const functions = {
     exitFullscreen,
     setInputFieldText,
     writeInputToFile,
-    print
+    print,
+    installFirefoxExtension
 }
 
 ipcRenderer.on("action", (_, name, ...args) => {