+++ app/js/suggest.js
@@ -18,6 +18,7 @@
 /* global COMMAND CONTEXTMENU HISTORY INPUT MODES SETTINGS TABS UTIL */
 "use strict"
 
+const {ipcRenderer} = require("electron")
 const path = require("path")
 const fs = require("fs")
 
@@ -267,8 +268,27 @@ const suggestCommand = search => {
         suggestFiles(location).forEach(l => addCommand(`write ${l.path}`))
     }
     // Command: mkviebrc
-    if ("mkviebrc full".startsWith(search) && !confirm) {
+    if ("mkviebrc full".startsWith(search)) {
         addCommand("mkviebrc full")
+    }
+    // Command: extensions
+    if ("extensions".startsWith(command) && !confirm && args.length < 3) {
+        if (args.length < 2) {
+            for (const action of ["install", "list", "remove"]) {
+                if (action.startsWith(args[0] || "") && action !== args[0]) {
+                    addCommand(`extensions ${action}`)
+                }
+            }
+        }
+        if (args.length >= 1) {
+            ipcRenderer.sendSync("list-extensions").forEach(e => {
+                const id = e.path.replace(/\/$/g, "").replace(/^.*\//g, "")
+                if (`remove ${id}`.startsWith(args.join(" "))) {
+                    addCommand(`extensions remove ${id}`,
+                        `${e.name}: ${e.version}`)
+                }
+            })
+        }
     }
     // Command: call
     if ("call".startsWith(command) && !confirm) {